Safari-609.1.11.1/LayoutTests/svg/animations/force-use-shadow-tree-recreation-while-animating-expected.txt

SVG 1.1 dynamic animation tests


This test forces use shadow tree recreation while an animating is running

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.width.animVal.value is 10
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 55
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 55
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 10
PASS successfullyParsed is true

TEST COMPLETE
